Move test cases for routes to split files and add missing content-negotiation test dependencies.
